feat(namespacing): updated architecture to better facilitate namespacing
rather then generating a single `pod-names.js` file with all of the
namespaced styles, moving to a mechanism where each style file
set (style files that are shared the same namespace at the same depth)
gets it's own `style-info.js` file that has the namespacing for that
component.

For components with a `component.js` file, using a babel plugin to
import the `style-info.js` and add the styleNamespace as a class
property, as well as adding it to the classNameBindings.

For routing styles, using a lookup pattern where we are looking up the
given namespace through the owner lookup.

For components without a component.js, creating a base one that defines
the import and sets the class with the layout.

For template-only-glimmer-components, wrapping the contents of the
template in a `let` block that defines the `styleNamespace`.

* moving environment.js 'ember-component-css' config values to
emberCliStyles values that are set in the ember-cli-build.js file.

* Also deprecating the `pod-styles` name for the imported style file
in favor of an `ember-styles` named file.

* Also migrating things to their latest versions as well and moving
things to proper classes.

* Also using preprocessor registries rather then the `treeFor` hooks.

* Also migrated the tests to use the latest modules and hooks.
Setup some base acceptance tests that the different style types share.